+++ cli/src/main/scala/com/github/gchudnov/kprojekt/runner/Runner.scala
@@ -1,0 +1,11 @@
+package com.github.gchudnov.kprojekt.runner
+
+import com.github.gchudnov.kprojekt.AppConfig
+
+object Runner {
+
+  def run(conf: AppConfig): Unit = {
+
+  }
+
+}
